requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.4
python-dateutil==2.8.2
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.3
